+++ build.sh
@@ -1,3 +1,5 @@
 #!/bin/sh
 
-javac -classpath . *.java
+javac --source-path ./corestructs/ -d . ./corestructs/*.java
+jar cf ./lib/corestructs.jar ./corestructs/*.class
+find ./corestructs/ -type f -name '*.class' -delete
