+++ corestructs/IOneDimensionStorable.java
@@ -30,7 +30,11 @@ OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 */
 
-interface ITwoDimensionIterable<K,V> {
-    int getKeyIndex(Object key);
-    int getValueIndex(Object val);
+package corestructs;
+
+interface IOneDimensionStorable<T> {
+    T get(int index);
+    void add(Object toBeAdded);
+    void replace(int index, Object toBeReplaced);
+    void replaceAll(Object toBeReplaced);
 }
