+++ corestructs/ISortable.java
@@ -30,6 +30,8 @@ OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 */
 
-interface IOneDimensionIterable<T> {
-    int getIndex(Object object);
+package corestructs;
+
+interface ISortable<T> {
+    void sort();
 }
