+++ corestructs/ITwoDimensionStorable.java
@@ -30,28 +30,12 @@ OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 */
 
-public class DynamicUniqueStore<T> extends DynamicStore<T> {
-    public DynamicUniqueStore() {
-        super();
-    }
-    protected void setItem(Item<?> newItem) {
-        try {
-            super.getIndex(newItem);
-        } catch (Exception ex) {
-            Item<?>[] newItems = null;
-            int currentIndex = super.getCurrentIndex();
-            Item<?>[] items = super.getItems();
-            int itemsLength = items.length;
-            newItems = new Item[itemsLength + 1];
-            for (int i = 0; i < itemsLength; i++) {
-                Item<?> currentItem = items[i];
-                if (currentItem != null) {
-                    newItems[i] = currentItem;
-                }
-            }
-            newItems[currentIndex] = newItem;
-            super.setItems(newItems);
-            super.setCurrentIndex(currentIndex + 1);
-        }
-    }
+package corestructs;
+
+interface ITwoDimensionStorable<K,V> {
+    K getKeyAtPosition(int index);
+    V getValueAtPosition(int index);
+    K getKey(Object val);
+    V getValue(Object key);
+    void add(Object key, Object value);
 }
