+++ corestructs/ItemStore.java
@@ -30,6 +30,8 @@ OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 */
 
+package corestructs;
+
 abstract class ItemStore<T> implements ISortable<T> {
     private int currentIndex;
     private Item<?>[] items;
